- Added extra score orb to tutorial level 1

+++ Game_Culminating_MansivA/TutorialPart2.cs
@@ -144,7 +144,7 @@ namespace Game_Culminating_MansivA
         {
             // makes the player jump by using a quadratic function for a parabola, Math.Sign for falling down, and a coefficent to have the player "float" near the top
             // of their jump
-            intJumpVelocity = (int)(0.2 * Math.Pow((double)intJumpPower, 2.0) * Math.Sign(intJumpPower));
+            intJumpVelocity = (int)(0.2 * Math.Pow(intJumpPower, 2.0) * Math.Sign(intJumpPower));
         }
         // Handles the players jump
         private void playerJump()
